delete icon displays next to food rows
