+++ index.html
@@ -3,6 +3,7 @@
 <head>
   <meta charset="utf-8">
   <title>Quantified Self - Diary</title>
+  <link rel="stylesheet" href="path/to/font-awesome/css/font-awesome.min.css">
   <style>
       table {
       color: #333;
@@ -31,6 +32,15 @@
     .green {
       color: green
     }
+
+    .delete-icon {
+      height: 15px
+    }
+
+    .fa-minus-circle {
+      color: red;
+    }
+
     /*table {
         font-family: arial, sans-serif;
         border-collapse: collapse;
@@ -111,7 +121,7 @@
 
     </tbody>
     <tfoot class='meal-sub-total'>
-      
+
     </tfoot>
   </table>
