+++ lib/ajax_requests/meal.js
@@ -1,6 +1,6 @@
 const $ = require('jquery')
 const API = "https://rocky-earth-59921.herokuapp.com";
-
+const mainMealResponse = require("../response-handlers/meal-responses")
 
 $(document).ready(function() {
   getAllMeals()
@@ -8,29 +8,14 @@ $(document).ready(function() {
   addFoodToMeal()
 })
 
-const goalKey = {
-  Breakfast: 400,
-  Lunch: 600,
-  Dinner: 800,
-  Snack: 200
-}
+
 
 function getAllMeals() {
   $.ajax({
     url: API + '/api/v1/meals',
     method: 'GET',
-  }).then(function(data) {
-    data.forEach(function(meal) {
-      var sum = 0
-      meal.foods.forEach(function(food) {
-        sum += food.calories
-        $(`#${meal.name.toLowerCase()} #foods-table-body`).append('<tr><td class="food-name-cell">' + food.name + '</td><td class="calorie-cell">' + food.calories + '</td></tr>');
-      })
-      // $(`#${meal.name.toLowerCase()}`).children('tfoot').empty()
-
-      $(`#${meal.name.toLowerCase()}`).children('tfoot').append('<tr id="totalColumn"><td class="totalcal" id="breakfast">Total Calories:</td><td class="meal-totals">' + sum + '</td></tr>');
-      $(`#${meal.name.toLowerCase()}`).children('tfoot').append(`<tr id="remainingColumn"><td class="remainingCals ${meal.name.toLowerCase()}"> Remaining Calories:</td><td class="remaining-totals"> ${goalKey[meal.name] - sum} </td></tr>`);
-    })
+  }).then(function (data) {
+      mainMealResponse.mainMealResponse(data);
       displayAddButtons(data);
       renderRemainingColor();
       totalsTable();
