+++ lib/index.js
@@ -7,3 +7,4 @@ var someMeal = new Meal();
 const bs = require('./ajax_requests/food');
 
 const foodDiary = require('./ajax_requests/meal');
+const mealResponse = require('./response-handlers/meal-responses');
